requirements_dev.txt
build==0.10.0

black==23.3.0
flake8==7.1.0
isort==5.12.0
mypy==1.1.1
pylint==3.2.5
pytest==7.2.2
python-dotenv==1.0.1
